pyproject.toml
[tool.poetry]
name = "djangotutorials"
version = "0.1.0"
description = ""
authors = ["adrianfhr <adrianfahriaffandi@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
django-filter = "^24.3"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
